Implement support for a deadman's switch.
Deadman's switches are alerts that are always firing. This way, we can check
that the alerting stack is functional. We never want to receive an alert for
a deadman's switch, but we want to receive an alert when it is missing.

This commit adds these configuration options:

- deadman_switch_name: the name of the deadman's switch alert.
- deadman_switch_interval: how often the alert is expected to arrive. Set this
  a bit higher than how often the alert is sent to prevent false positives.
- deadman_switch_alert_interval: how often you want the bot to alert you of
  prolonged missing deadman's switch alerts.
- deadman_switch_alert_chat_id: to what chat you want the alert to be sent.

Alerts with the deadman_switch_name are removed from incoming posts.  If the
alert is the only one firing in a group, the overall status of the group is
overwritten to "resolved".